Separate auto-connection and best contest picking to new API
It was bugging me that RESTContestSource is an API class, but has been outputting messages to the console and auto-connecting while it validates the URL. For some uses you may not want auto-connect or console output. Separated out a new ContestAPIHelper class that does auto-connection/picking best contest; clients who need this function can use it before creating a RESTContestSource.

Also fixed an issue with connecting to a root URL like /api (fixed child URLs) and made minor tweaks to messages.
